Add scribe1 target using hf 5.16.3 firmware
The default `kindlehf` target is missing `ace` libraries which I need
for Bluetooth hacking. My guess it's because it uses an old Paperwhite
with no BT support for the `kindlehf` target.

Instead of replacing it and possibly breaking people's workflows, I
just decided to add a new target of a device I know supports BT and uses
HF.