Add automated builds for Docker images
With a bit of elbow grease from Github Actions we can automatically
build Docker images for a given device/firmware combination. With this
setup, you can push a tag with this format `device-firmware_version` and
the CI will take care of building it and pushing it to DockerHub/Gihub
Registry.