Use @DIRNAME@ in meson-crosscompile.txt
This makes it so the paths to the different compiler binaries,
pkg-config, as well as the sysroot, are not hardcoded to the environment
that built them, and instead they are relative to the directory that
houses the crosscompile file. This effectively makes it so, once we
build the SDK for a given device/version, you can just export it,
similarly to how the koxtoolchain does it.