Always use the latest koxtoolchain release
Github offers these helpful URLs that always link to the latest release.
It's a bit annoying that I then need to resolve it for the proper link
in the release notes (otherwise it would lead you to the latest release
_when_ you click it, not from when it was built), but otherwise pretty
handy solution. With this I no longer need to worry about updating this
manually.